chore: refactor main code into functions
